phase2/backend/requirements.txt
fastapi>=0.115,<0.120
uvicorn[standard]>=0.30,<0.40
python-multipart>=0.0.9,<0.1
pydantic>=2.6,<3
pydantic-settings>=2.2,<3
openai>=1.54,<2
httpx>=0.27,<0.29
beautifulsoup4>=4.12,<5
pandas>=2.2,<3
gitpython>=3.1,<4
watchfiles>=0.24,<1
pypdf>=4.3,<6
